requirements.txt
# Web
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# PyTorch CPU
torch==2.2.1
torchvision==0.17.1

# ML / scientific
numpy==1.26.4
scipy==1.13.1
pandas==2.1.4
scikit-learn==1.4.2
joblib==1.4.2
matplotlib==3.8.4

# Image
Pillow==10.3.0
opencv-python-headless

# PDF
jinja2==3.1.4
reportlab==4.1.0
weasyprint==62.3
# WeasyPrint 62.x breaks with pydyf >= 0.11.0 during PDF generation.
pydyf==0.10.0

# Segmentation
albumentations==1.3.1
albucore==0.0.3
segmentation-models-pytorch==0.3.3
tifffile==2023.7.10
